Update scripts and UI

+++ setup.sh
@@ -1,20 +1,9 @@
 #!/bin/bash
 set -euo pipefail
 
-# Install dirmngr
-echo "# Install 'dirmngr'..."
-[[ $(apt list --installed dirmngr 2>/dev/null | grep dirmngr) ]] || apt-get update
-apt-get install -y dirmngr
-
-# Add some package list
-echo "# Add some package list..."
-#curl -fsSL https://packages.redis.io/gpg | gpg --yes --dearmor -o /usr/share/keyrings/redis-archive-keyring.gpg
-#echo "deb [signed-by=/usr/share/keyrings/redis-archive-keyring.gpg] https://packages.redis.io/deb $(lsb_release -cs) main" | tee /etc/apt/sources.list.d/redis.list
-#[[ $(apt-key list --list-keys 8B48AD6246925553 2>/dev/null) ]] || apt-key adv --keyserver keyserver.ubuntu.com --recv-keys 8B48AD6246925553
-[[ $(apt-key list --list-keys 648ACFD622F3D138 2>/dev/null) ]] || apt-key adv --keyserver keyserver.ubuntu.com --recv-keys 648ACFD622F3D138
-[[ $(apt-key list --list-keys 0E98404D386FA1D9 2>/dev/null) ]] || apt-key adv --keyserver keyserver.ubuntu.com --recv-keys 0E98404D386FA1D9
-echo "deb http://deb.debian.org/debian/ unstable main" | tee /etc/apt/sources.list.d/unstable.list
-printf 'Package: *\nPin: release a=unstable\nPin-Priority: 150\n' | tee /etc/apt/preferences.d/limit-unstable
+# Configure update to the crone
+echo "# Configure crone update..."
+ln -sf /opt/opcb-release/update.sh /etc/cron.daily/opcb-update
 
 # Installing some package's
 # + hostapd
@@ -23,6 +12,8 @@ printf 'Package: *\nPin: release a=unstable\nPin-Priority: 150\n' | tee /etc/apt
 # + network-manager
 # + wireguard
 # + nginx
+# + qrencode
+# + imagemagick
 echo "# Installing some package's..."
 apt-get update
 apt-get install -y hostapd dnsmasq resolvconf network-manager wireguard nginx qrencode imagemagick
@@ -89,21 +80,21 @@ else
     echo "Skip 'fix-mac-addr', files  not found."
 fi
 
-# Configure update to the crone
-echo "# Configure crone update..."
-ln -sf /opt/opcb-release/update.sh /etc/cron.daily/opcb-update
-
 # Configure wireguard
 echo "# Configure WireGuard VPN..."
+echo "Install wg-watchdog to the cron..."
+ln -sf /opt/opcb-release/wireguard/wg-watchdog.cron /etc/cron.d/wg-watchdog
+echo "Start wireguard VPN service..."
 file1="/opt/opcb-release/wireguard/wg0.conf"
 file2="/etc/wireguard/wg0.conf"
-if [[ -f "$file1" ]]; then
+if [[ -f "$file1" ]] && [[ "$(cat $file1)" != "" ]]; then
     ln -sf "$file1" "$file2"
     systemctl restart wg-quick@wg0
     systemctl enable wg-quick@wg0
-    # Install wg-watchdog to the cron.d
-    ln -sf /opt/opcb-release/wireguard/wg-watchdog.cron /etc/cron.d/wg-watchdog
+else
+    echo "Skip but 'wg0.conf' is empty or not found."
 fi
+
 
 # Configure Nginx
 echo "# Configure Nginx..."
